+++ v2/Source/Controllers/Pin/Drivers/VNI8200XP/eosPinDriver_VNI8200XP.cpp
@@ -10,7 +10,7 @@ using namespace htl;
 
 PinDriver_VNI8200XP::PinDriver_VNI8200XP(
 	Device_VNI8200XP *dev,
-	uint8_t pinNumber):
+	unsigned pinNumber):
 
 	_dev {dev},
 	_pinMask {uint8_t(1 << pinNumber)} {
